Cassa comune

+++ app/routes.py
@@ -1,13 +1,11 @@
 # -*- coding: utf8 -*-
-import os
-import secrets
-from PIL import Image
 from flask import render_template, url_for, redirect, request, flash
 from grouppy import app
-from forms import LoginForm, RegisterForm, AddFriendForm, AddTripForm, UserSettingsForm, EditFriendForm
-from models import User, Friend, Trip
+from forms import LoginForm, RegisterForm, AddFriendForm, AddTripForm, UserSettingsForm, EditFriendForm, TransazioneForm
+from models import User, Friend, Trip, Transazione
 from flask_login import login_user, login_required, logout_user, current_user
 from werkzeug.security import generate_password_hash
+from datetime import datetime
 
 from google.appengine.ext import blobstore, ndb
 from google.appengine.api import images
@@ -65,11 +63,14 @@ def dashboard():
         ancestor=current_user.key).order(-Friend.score).fetch()
     best_friends = friends[0:2]
     worst_friends = friends[-2:]
+    transazioni = Transazione.query(
+        ancestor=current_user.key).order(-Transazione.data).fetch()
     return render_template('dashboard.html', user=current_user, friends=friends,
-                           best_friends=best_friends, worst_friends=worst_friends)
+                           best_friends=best_friends, worst_friends=worst_friends,
+                           transazioni=transazioni)
 
 
-@app.route('/add_friend', methods=['GET', 'POST'])
+@app.route('/friend/add', methods=['GET', 'POST'])
 @login_required
 def add_friend():
     form = AddFriendForm()
@@ -130,16 +131,18 @@ def friend_edit(friend_id):
         friend.nome = form.nome.data
         friend.cognome = form.cognome.data
 
-        uploaded_file = request.files.get('immagine')
-        file_headers = uploaded_file.headers['Content-Type']
-        blob_key = werkzeug.parse_options_header(file_headers)[1]['blob-key']
-        image_url = images.get_serving_url(blob_key=blob_key)
+        if form.immagine.data:
+            uploaded_file = request.files.get('immagine')
+            file_headers = uploaded_file.headers['Content-Type']
+            blob_key = werkzeug.parse_options_header(file_headers)[
+                1]['blob-key']
+            image_url = images.get_serving_url(blob_key=blob_key)
 
-        if friend.immagine_blob_key:
-            blobstore.BlobInfo.get(friend.immagine_blob_key).delete()
+            if friend.immagine_blob_key:
+                blobstore.BlobInfo.get(friend.immagine_blob_key).delete()
 
-        friend.immagine_url = image_url
-        friend.immagine_blob_key = blobstore.BlobKey(blob_key)
+            friend.immagine_url = image_url
+            friend.immagine_blob_key = blobstore.BlobKey(blob_key)
 
         friend.put()
         message = friend.nome + " " + friend.cognome + " modificato con successo!"
@@ -163,5 +166,72 @@ def friend_delete(friend_id):
     if friend.immagine_blob_key:
         blobstore.BlobInfo.get(friend.immagine_blob_key).delete()
     friend.key.delete()
+    flash(message, 'success')
+    return redirect(url_for('dashboard'))
+
+
+@app.route('/add_transazione', methods=['GET', 'POST'])
+def add_transazione():
+    form = TransazioneForm()
+    if form.validate_on_submit():
+        transazione = Transazione(titolo=form.titolo.data,
+                                  descrizione=form.descrizione.data,
+                                  data=form.data.data,
+                                  costo=form.costo.data,
+                                  parent=current_user.key)
+        transazione.put()
+        current_user.cassa = current_user.cassa + form.costo.data
+        current_user.put()
+        message = 'Transazione creata con successo!'
+        flash(message, 'success')
+        return redirect(url_for('dashboard'))
+    return render_template('transazione_form.html',
+                           form=form,
+                           titolo='Aggiungi transazione',
+                           dest=url_for('add_transazione'))
+
+
+@app.route('/transazione/edit/<transazione_id>', methods=['GET', 'POST'])
+def transazione_edit(transazione_id):
+    transazione = Transazione.get_by_id(
+        int(transazione_id), parent=current_user.key)
+    if not transazione:
+        return render_template('_404.html', message='Transazione non esistente'), 404
+    form = TransazioneForm()
+    if request.method == 'POST' and form.validate_on_submit():
+        current_user.cassa = current_user.cassa - transazione.costo
+        transazione.titolo = form.titolo.data
+        transazione.data = form.data.data
+        transazione.costo = form.costo.data
+        transazione.descrizione = form.descrizione.data
+        current_user.cassa = current_user.cassa + transazione.costo
+        transazione.put()
+        current_user.put()
+        message = 'Transazione modificata con successo!'
+        flash(message, 'success')
+        return redirect(url_for('dashboard'))
+    elif request.method == 'GET':
+        form.titolo.data = transazione.titolo
+        form.data.data = transazione.data
+        form.costo.data = transazione.costo
+        form.descrizione.data = transazione.descrizione
+    return render_template('transazione_form.html',
+                           form=form,
+                           titolo='Modifica transazione',
+                           dest=url_for('transazione_edit',
+                                        transazione_id=transazione_id),
+                           transazione=transazione)
+
+
+@app.route('/transazione/delete/<transazione_id>')
+def transazione_delete(transazione_id):
+    transazione = Transazione.get_by_id(
+        int(transazione_id), parent=current_user.key)
+    if not transazione:
+        return render_template('_404.html', message='Transazioen non esistente'), 404
+    message = "Transazione eliminata con successo!"
+    current_user.cassa -= transazione.costo
+    current_user.put()
+    transazione.key.delete()
     flash(message, 'success')
     return redirect(url_for('dashboard'))
